detangle logic for fetching select 2 options for Users, CaseList, Locations, etc filters from the view so that it can be re-used in other views utilizing a seperate structure / hierarchy
