+++ corehq/apps/reports/filters/location.py
@@ -2,6 +2,9 @@ from __future__ import absolute_import
 from __future__ import unicode_literals
 from memoized import memoized
 from corehq.apps.locations.models import SQLLocation
+from corehq.apps.reports.filters.controllers import (
+    LocationGroupOptionsController,
+)
 from .users import ExpandedMobileWorkerFilter
 from .api import EmwfOptionsView
 from django.urls import reverse
@@ -53,8 +56,6 @@ class LocationGroupFilter(ExpandedMobileWorkerFilter):
 class LocationGroupFilterOptions(EmwfOptionsView):
 
     @property
-    def data_sources(self):
-        return [
-            (self.get_groups_size, self.get_groups),
-            (self.get_locations_size, self.get_locations),
-        ]
+    @memoized
+    def options_controller(self):
+        return LocationGroupOptionsController(self.request, self.domain, self.search)
